+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `Tornex.Scheduler.Timer` failing to dump queries in other nodes
 
 ### Changed
-- Changed `Tornex.Scheduler.Bucket.dump/1` to cast instead of callthe `:dump` request to the GenServer
+- Changed `Tornex.Scheduler.Bucket.dump/1` to cast instead of call the `:dump` request to the GenServer
 
 ## [0.6.0] - 2026-06-05
 ### Added